+++ src/main/java/io/games/play_book/player_participation/model/PlayerParticipationDTO.java
@@ -10,7 +10,7 @@ import lombok.Setter;
 @Setter
 public class PlayerParticipationDTO {
 
-  private Integer participationId;
+  private Long participationId;
 
   @NotNull private Boolean participated;
